refactor: 💡 Change chart label visualization
